+++ inc/Constants.php
@@ -21,6 +21,4 @@ class Constants extends LibConstants
     public const CACHE_PATH = __DIR__ . '/../.cache';
     public const CONFIG_PATH = __DIR__ . '/standalone-settings.php';
     public const VERSION = 'v2.3.1';
-    public const GEOLOCATION_DIR = __DIR__ . '/../geolocation';
-    public const TLS_DIR = __DIR__ . '/../tls';
 }